--- src/main/java/Duke/Commands/CheckCommand.java
@@ -1,24 +1,0 @@
-package Duke.Commands;
-
-import Duke.Tools.Storage;
-import Duke.Tools.TaskList;
-import Duke.Tools.Ui;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class CheckCommand extends Command {
-
-    private final DateTimeFormatter formatter;
-
-    public CheckCommand(String fullCommand, DateTimeFormatter formatter) {
-        super(fullCommand);
-        this.formatter = formatter;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        String content = this.fullCommand.replaceAll("^\\s*check\\s*", "");
-        ui.printScheduledTasks(tasks, LocalDateTime.parse(content, this.formatter));
-    }
-}
