--- src/main/java/Duke/Exceptions/BackwardsTimeException.java
@@ -1,9 +1,0 @@
-package Duke.Exceptions;
-
-public class BackwardsTimeException extends Exception {
-
-    public BackwardsTimeException() {
-        super("Sorry... The provided timeframe is impossible as the deadline is earlier than the start...");
-    }
-
-}
