--- src/main/java/Duke/Exceptions/EmptyTaskListException.java
@@ -1,13 +1,0 @@
-package Duke.Exceptions;
-
-/**
- * Represents an Duke.Exceptions.EmptyTaskListException.
- *
- * @author Rayson
- */
-public class EmptyTaskListException extends Exception {
-
-    public EmptyTaskListException() {
-        super("☹ OOPS!!! The task list cannot be empty.");
-    }
-}
