--- src/main/java/Duke/Exceptions/NoIndexException.java
@@ -1,8 +1,0 @@
-package Duke.Exceptions;
-
-public class NoIndexException extends Exception {
-
-    public NoIndexException(String message) {
-        super(String.format("☹ OOPS!!! The index %s does not exist.", message));
-    }
-}
