--- src/main/java/Duke/Exceptions/UnrecognisedFormatException.java
@@ -1,9 +1,0 @@
-package Duke.Exceptions;
-
-public class UnrecognisedFormatException extends Exception {
-
-    public UnrecognisedFormatException() {
-        super("Sorry!!! This data file is corrupted and we do not recognise the format of the file.");
-    }
-
-}
