+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,0 +1,19 @@
+package duke.commands;
+
+import duke.tools.Storage;
+import duke.tools.TaskList;
+import duke.tools.Ui;
+
+public class ExitCommand extends Command {
+
+    public ExitCommand() {
+        super("");
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        this.isExit = true;
+        storage.writeFile(tasks.retrieveForStorage());
+        ui.showOutro();
+    }
+}
