+++ src/main/java/duke/commands/ListCommand.java
@@ -1,0 +1,21 @@
+package duke.commands;
+
+import duke.exceptions.EmptyTaskListException;
+import duke.tools.Storage;
+import duke.tools.TaskList;
+import duke.tools.Ui;
+
+public class ListCommand extends Command {
+
+    public ListCommand() {
+        super("");
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws EmptyTaskListException {
+        if (tasks.size() == 0) {
+            throw new EmptyTaskListException();
+        }
+        ui.printTasks(tasks);
+    }
+}
