+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,0 +1,30 @@
+package duke.commands;
+
+import duke.exceptions.NoIndexException;
+import duke.tools.Storage;
+import duke.tasks.Task;
+import duke.tools.TaskList;
+import duke.tools.Ui;
+
+public class MarkCommand extends Command {
+
+    public MarkCommand(String fullCommand) {
+        super(fullCommand);
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws NoIndexException {
+        String desc = this.fullCommand.replaceAll("[^0-9]", "");
+        if (desc.equals("")) {
+            throw new NoIndexException("No Index");
+        }
+        int index = Integer.parseInt(desc);
+        if (tasks.size() < index || index < 1) {
+            throw new NoIndexException(Integer.toString(index));
+        }
+
+        Task task = tasks.get(index - 1);
+        task.markDone();
+        ui.showMarkMessage(task);
+    }
+}
