+++ src/main/java/duke/exceptions/BackwardsTimeException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class BackwardsTimeException extends Exception {
+
+    public BackwardsTimeException() {
+        super("Sorry... The provided timeframe is impossible as the deadline is earlier than the start...");
+    }
+
+}
