+++ src/main/java/duke/exceptions/EmptyTaskListException.java
@@ -1,0 +1,13 @@
+package duke.exceptions;
+
+/**
+ * Represents an Duke.Exceptions.EmptyTaskListException.
+ *
+ * @author Rayson
+ */
+public class EmptyTaskListException extends Exception {
+
+    public EmptyTaskListException() {
+        super("☹ OOPS!!! The task list cannot be empty.");
+    }
+}
