+++ src/main/java/duke/exceptions/NoDescriptionException.java
@@ -1,0 +1,14 @@
+package duke.exceptions;
+
+/**
+ * Represents a Duke.Exceptions.NoDescriptionException.
+ *
+ * @author Rayson
+ */
+public class NoDescriptionException extends Exception {
+
+    public NoDescriptionException(String message) {
+        super(String.format("☹ OOPS!!! Where is your description for %s?", message));
+    }
+
+}
