+++ src/main/java/duke/exceptions/UnrecognisedFormatException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class UnrecognisedFormatException extends Exception {
+
+    public UnrecognisedFormatException() {
+        super("Sorry!!! This data file is corrupted and we do not recognise the format of the file.");
+    }
+
+}
